+++ lib/lanttern/grades_reports.ex
@@ -4,18 +4,373 @@ defmodule Lanttern.GradesReports do
   """
 
   import Ecto.Query, warn: false
-  alias Lanttern.Reporting
   alias Lanttern.Schools.Student
   alias Lanttern.Repo
   import Lanttern.RepoHelpers
 
   alias Lanttern.Assessments.AssessmentPointEntry
+  alias Lanttern.GradesReports.GradesReport
+  alias Lanttern.GradesReports.GradesReportCycle
+  alias Lanttern.GradesReports.GradesReportSubject
   alias Lanttern.GradesReports.StudentGradeReportEntry
   alias Lanttern.Grading
   alias Lanttern.Grading.OrdinalValue
-  alias Lanttern.Reporting.GradesReportCycle
-  alias Lanttern.Reporting.GradesReportSubject
   alias Lanttern.Reporting.StudentReportCard
+
+  @doc """
+  Returns the list of grade reports.
+
+  ## Options
+
+  - `:preloads` – preloads associated data
+  - `:load_grid` – (bool) preloads school cycle and grades report cycles/subjects (with school cycle/subject preloaded)
+
+  ## Examples
+
+      iex> list_grades_reports()
+      [%GradesReport{}, ...]
+
+  """
+  def list_grades_reports(opts \\ []) do
+    GradesReport
+    |> apply_list_grades_reports_opts(opts)
+    |> Repo.all()
+    |> maybe_preload(opts)
+  end
+
+  defp apply_list_grades_reports_opts(queryable, []), do: queryable
+
+  defp apply_list_grades_reports_opts(queryable, [{:load_grid, true} | opts]),
+    do: apply_list_grades_reports_opts(grid_query(queryable), opts)
+
+  defp apply_list_grades_reports_opts(queryable, [_ | opts]),
+    do: apply_list_grades_reports_opts(queryable, opts)
+
+  defp grid_query(queryable) do
+    from(
+      gr in queryable,
+      join: sc in assoc(gr, :school_cycle),
+      left_join: grc in assoc(gr, :grades_report_cycles),
+      left_join: grc_sc in assoc(grc, :school_cycle),
+      left_join: grs in assoc(gr, :grades_report_subjects),
+      left_join: grs_s in assoc(grs, :subject),
+      order_by: [asc: grc_sc.end_at, desc: grc_sc.start_at, asc: grs.position],
+      preload: [
+        school_cycle: sc,
+        grades_report_cycles: {grc, [school_cycle: grc_sc]},
+        grades_report_subjects: {grs, [subject: grs_s]}
+      ]
+    )
+  end
+
+  @doc """
+  Gets a single grade report.
+
+  Returns `nil` if the grade report does not exist.
+
+  ## Options:
+
+  - `:preloads` – preloads associated data
+  - `:load_grid` – (bool) preloads school cycle and grades report cycles/subjects (with school cycle/subject preloaded)
+
+  ## Examples
+
+      iex> get_grades_report!(123)
+      %GradesReport{}
+
+      iex> get_grades_report!(456)
+      nil
+
+  """
+  def get_grades_report(id, opts \\ []) do
+    GradesReport
+    |> apply_get_grades_report_opts(opts)
+    |> Repo.get(id)
+    |> maybe_preload(opts)
+  end
+
+  defp apply_get_grades_report_opts(queryable, []), do: queryable
+
+  defp apply_get_grades_report_opts(queryable, [{:load_grid, true} | opts]),
+    do: apply_get_grades_report_opts(grid_query(queryable), opts)
+
+  defp apply_get_grades_report_opts(queryable, [_ | opts]),
+    do: apply_get_grades_report_opts(queryable, opts)
+
+  @doc """
+  Gets a single grade report.
+
+  Same as `get_grades_report/2`, but raises `Ecto.NoResultsError` if the grade report does not exist.
+
+  ## Examples
+
+      iex> get_grades_report!(123)
+      %GradesReport{}
+
+      iex> get_grades_report!(456)
+      ** (Ecto.NoResultsError)
+
+  """
+  def get_grades_report!(id, opts \\ []) do
+    GradesReport
+    |> apply_get_grades_report_opts(opts)
+    |> Repo.get!(id)
+    |> maybe_preload(opts)
+  end
+
+  @doc """
+  Creates a grade report.
+
+  ## Examples
+
+      iex> create_grades_report(%{field: value})
+      {:ok, %GradesReport{}}
+
+      iex> create_grades_report(%{field: bad_value})
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def create_grades_report(attrs \\ %{}) do
+    %GradesReport{}
+    |> GradesReport.changeset(attrs)
+    |> Repo.insert()
+  end
+
+  @doc """
+  Updates a grade report.
+
+  ## Examples
+
+      iex> update_grades_report(grades_report, %{field: new_value})
+      {:ok, %ReportCard{}}
+
+      iex> update_grades_report(grades_report, %{field: bad_value})
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def update_grades_report(%GradesReport{} = grades_report, attrs) do
+    grades_report
+    |> GradesReport.changeset(attrs)
+    |> Repo.update()
+  end
+
+  @doc """
+  Deletes a grade report.
+
+  ## Examples
+
+      iex> delete_grades_report(grades_report)
+      {:ok, %ReportCard{}}
+
+      iex> delete_grades_report(grades_report)
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def delete_grades_report(%GradesReport{} = grades_report) do
+    Repo.delete(grades_report)
+  end
+
+  @doc """
+  Returns an `%Ecto.Changeset{}` for tracking grade report changes.
+
+  ## Examples
+
+      iex> change_grades_report(grades_report)
+      %Ecto.Changeset{data: %ReportCard{}}
+
+  """
+  def change_grades_report(%GradesReport{} = grades_report, attrs \\ %{}) do
+    GradesReport.changeset(grades_report, attrs)
+  end
+
+  @doc """
+  Returns the list of grades report subjects.
+
+  Results are ordered by position and preloaded subjects.
+
+  ## Examples
+
+      iex> list_grades_report_subjects(1)
+      [%GradesReportSubject{}, ...]
+
+  """
+  @spec list_grades_report_subjects(grades_report_id :: integer()) :: [
+          GradesReportSubject.t()
+        ]
+
+  def list_grades_report_subjects(grades_report_id) do
+    from(grs in GradesReportSubject,
+      order_by: grs.position,
+      join: s in assoc(grs, :subject),
+      preload: [subject: s],
+      where: grs.grades_report_id == ^grades_report_id
+    )
+    |> Repo.all()
+  end
+
+  @doc """
+  Add a subject to a grades report.
+
+  Result has subject preloaded.
+
+  ## Examples
+
+      iex> add_subject_to_grades_report(%{field: value})
+      {:ok, %GradesReportSubject{}}
+
+      iex> add_subject_to_grades_report(%{field: bad_value})
+      {:error, %Ecto.Changeset{}}
+  """
+
+  @spec add_subject_to_grades_report(map()) ::
+          {:ok, GradesReportSubject.t()} | {:error, Ecto.Changeset.t()}
+
+  def add_subject_to_grades_report(attrs \\ %{}) do
+    %GradesReportSubject{}
+    |> GradesReportSubject.changeset(attrs)
+    |> set_grades_report_subject_position()
+    |> Repo.insert()
+    |> maybe_preload(preloads: :subject)
+  end
+
+  # skip if not valid
+  defp set_grades_report_subject_position(%Ecto.Changeset{valid?: false} = changeset),
+    do: changeset
+
+  # skip if changeset already has position change
+  defp set_grades_report_subject_position(
+         %Ecto.Changeset{changes: %{position: _position}} = changeset
+       ),
+       do: changeset
+
+  defp set_grades_report_subject_position(%Ecto.Changeset{} = changeset) do
+    grades_report_id =
+      Ecto.Changeset.get_field(changeset, :grades_report_id)
+
+    position =
+      from(
+        grs in GradesReportSubject,
+        where: grs.grades_report_id == ^grades_report_id,
+        select: grs.position,
+        order_by: [desc: grs.position],
+        limit: 1
+      )
+      |> Repo.one()
+      |> case do
+        nil -> 0
+        pos -> pos + 1
+      end
+
+    changeset
+    |> Ecto.Changeset.put_change(:position, position)
+  end
+
+  @doc """
+  Update grades report subjects positions based on ids list order.
+
+  ## Examples
+
+  iex> update_grades_report_subjects_positions([3, 2, 1])
+  :ok
+
+  """
+  @spec update_grades_report_subjects_positions([integer()]) :: :ok | {:error, String.t()}
+  def update_grades_report_subjects_positions(grades_report_subjects_ids),
+    do: update_positions(GradesReportSubject, grades_report_subjects_ids)
+
+  @doc """
+  Deletes a grades report subject.
+
+  ## Examples
+
+      iex> delete_grades_report_subject(grades_report_subject)
+      {:ok, %GradesReportSubject{}}
+
+      iex> delete_grades_report_subject(grades_report_subject)
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def delete_grades_report_subject(%GradesReportSubject{} = grades_report_subject),
+    do: Repo.delete(grades_report_subject)
+
+  @doc """
+  Returns the list of grades report cycles.
+
+  Results are ordered asc by cycle `end_at` and desc by cycle `start_at`, and have preloaded school cycles.
+
+  ## Examples
+
+  iex> list_grades_report_cycles(1)
+  [%GradesReportCycle{}, ...]
+
+  """
+  @spec list_grades_report_cycles(grades_report_id :: integer()) :: [
+          GradesReportCycle.t()
+        ]
+
+  def list_grades_report_cycles(grades_report_id) do
+    from(grc in GradesReportCycle,
+      join: sc in assoc(grc, :school_cycle),
+      preload: [school_cycle: sc],
+      where: grc.grades_report_id == ^grades_report_id,
+      order_by: [asc: sc.end_at, desc: sc.start_at]
+    )
+    |> Repo.all()
+  end
+
+  @doc """
+  Add a cycle to a grades report.
+
+  ## Examples
+
+    iex> add_cycle_to_grades_report(%{field: value})
+    {:ok, %GradesReportCycle{}}
+
+    iex> add_cycle_to_grades_report(%{field: bad_value})
+    {:error, %Ecto.Changeset{}}
+  """
+
+  @spec add_cycle_to_grades_report(map()) ::
+          {:ok, GradesReportCycle.t()} | {:error, Ecto.Changeset.t()}
+
+  def add_cycle_to_grades_report(attrs \\ %{}) do
+    %GradesReportCycle{}
+    |> GradesReportCycle.changeset(attrs)
+    |> Repo.insert()
+  end
+
+  @doc """
+  Updates a grades_report_cycle.
+
+  ## Examples
+
+      iex> update_grades_report_cycle(grades_report_cycle, %{field: new_value})
+      {:ok, %GradesReportCycle{}}
+
+      iex> update_grades_report_cycle(grades_report_cycle, %{field: bad_value})
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def update_grades_report_cycle(%GradesReportCycle{} = grades_report_cycle, attrs) do
+    grades_report_cycle
+    |> GradesReportCycle.changeset(attrs)
+    |> Repo.update()
+  end
+
+  @doc """
+  Deletes a grades report cycle.
+
+  ## Examples
+
+    iex> delete_grades_report_cycle(grades_report_cycle)
+    {:ok, %GradesReportCycle{}}
+
+    iex> delete_grades_report_cycle(grades_report_cycle)
+    {:error, %Ecto.Changeset{}}
+
+  """
+  def delete_grades_report_cycle(%GradesReportCycle{} = grades_report_cycle),
+    do: Repo.delete(grades_report_cycle)
 
   @doc """
   Returns the list of student_grade_report_entries.
@@ -148,7 +503,7 @@ defmodule Lanttern.GradesReports do
         grades_report_subject_id
       ) do
     # get grades report scale
-    %{scale: scale} = Reporting.get_grades_report!(grades_report_id, preloads: :scale)
+    %{scale: scale} = get_grades_report!(grades_report_id, preloads: :scale)
 
     from(
       e in AssessmentPointEntry,
@@ -322,7 +677,9 @@ defmodule Lanttern.GradesReports do
       scale: scale,
       grades_report_subjects: grades_report_subjects
     } =
-      Reporting.get_grades_report!(grades_report_id, preloads: [:scale, :grades_report_subjects])
+      get_grades_report!(grades_report_id,
+        preloads: [:scale, :grades_report_subjects]
+      )
 
     grades_report_subject_entries_grade_components =
       from(
@@ -441,7 +798,7 @@ defmodule Lanttern.GradesReports do
         grades_report_subject_id
       ) do
     # get grades report scale
-    %{scale: scale} = Reporting.get_grades_report!(grades_report_id, preloads: :scale)
+    %{scale: scale} = get_grades_report!(grades_report_id, preloads: :scale)
 
     students_entries_grade_components =
       from(
@@ -565,7 +922,9 @@ defmodule Lanttern.GradesReports do
       scale: scale,
       grades_report_subjects: grades_report_subjects
     } =
-      Reporting.get_grades_report!(grades_report_id, preloads: [:scale, :grades_report_subjects])
+      get_grades_report!(grades_report_id,
+        preloads: [:scale, :grades_report_subjects]
+      )
 
     students_grades_report_subject_entries_grade_components =
       from(
