+++ test/support/fixtures/reporting_fixtures.ex
@@ -55,53 +55,6 @@ defmodule Lanttern.ReportingFixtures do
     student_report_card
   end
 
-  @doc """
-  Generate a grade report.
-  """
-  def grades_report_fixture(attrs \\ %{}) do
-    {:ok, grades_report} =
-      attrs
-      |> Enum.into(%{
-        name: "some name",
-        info: "some info",
-        school_cycle_id: Lanttern.SchoolsFixtures.maybe_gen_cycle_id(attrs),
-        scale_id: Lanttern.GradingFixtures.maybe_gen_scale_id(attrs)
-      })
-      |> Lanttern.Reporting.create_grades_report()
-
-    grades_report
-  end
-
-  @doc """
-  Generate a grades_report_subject.
-  """
-  def grades_report_subject_fixture(attrs \\ %{}) do
-    {:ok, grades_report_subject} =
-      attrs
-      |> Enum.into(%{
-        grades_report_id: maybe_gen_grades_report_id(attrs),
-        subject_id: Lanttern.TaxonomyFixtures.maybe_gen_subject_id(attrs)
-      })
-      |> Lanttern.Reporting.add_subject_to_grades_report()
-
-    grades_report_subject
-  end
-
-  @doc """
-  Generate a grades_report_cycle.
-  """
-  def grades_report_cycle_fixture(attrs \\ %{}) do
-    {:ok, grades_report_cycle} =
-      attrs
-      |> Enum.into(%{
-        grades_report_id: maybe_gen_grades_report_id(attrs),
-        school_cycle_id: Lanttern.SchoolsFixtures.maybe_gen_cycle_id(attrs)
-      })
-      |> Lanttern.Reporting.add_cycle_to_grades_report()
-
-    grades_report_cycle
-  end
-
   # generator helpers
 
   def maybe_gen_report_card_id(%{report_card_id: report_card_id} = _attrs),
@@ -109,26 +62,4 @@ defmodule Lanttern.ReportingFixtures do
 
   def maybe_gen_report_card_id(_attrs),
     do: report_card_fixture().id
-
-  def maybe_gen_grades_report_id(%{grades_report_id: grades_report_id} = _attrs),
-    do: grades_report_id
-
-  def maybe_gen_grades_report_id(_attrs),
-    do: grades_report_fixture().id
-
-  def maybe_gen_grades_report_cycle_id(
-        %{grades_report_cycle_id: grades_report_cycle_id} = _attrs
-      ),
-      do: grades_report_cycle_id
-
-  def maybe_gen_grades_report_cycle_id(_attrs),
-    do: grades_report_cycle_fixture().id
-
-  def maybe_gen_grades_report_subject_id(
-        %{grades_report_subject_id: grades_report_subject_id} = _attrs
-      ),
-      do: grades_report_subject_id
-
-  def maybe_gen_grades_report_subject_id(_attrs),
-    do: grades_report_subject_fixture().id
 end
